agents/custom-agents/src/exchange_rate_lookup_agent/requirements.txt
# Exchange Rate Lookup Agent Dependencies
# HTTP client for API calls
httpx>=0.24.0

# Date parsing utilities
python-dateutil>=2.8.0

# Testing framework
pytest>=7.0.0

# Async testing support
pytest-asyncio>=0.21.0

# Type hints support
typing-extensions>=4.0.0

# Environment variable loading
python-dotenv>=1.0.0
